+++ ui/app/components/app/token-list/index.js
@@ -1,0 +1,1 @@
+export { default } from './token-list.container'
